+++ Hearthstone Deck Tracker/Hearthstone/CounterSystem/Counters/FriendlyAttacksThisGameCounter.cs
@@ -1,0 +1,52 @@
+﻿using System;
+using HearthDb.Enums;
+using Hearthstone_Deck_Tracker.Hearthstone.Entities;
+using Hearthstone_Deck_Tracker.LogReader.Interfaces;
+using Hearthstone_Deck_Tracker.Utility;
+
+namespace Hearthstone_Deck_Tracker.Hearthstone.CounterSystem.Counters;
+
+public class FriendlyAttacksThisGameCounter : NumericCounter
+{
+	protected override string? CardIdToShowInUI => HearthDb.CardIds.Collectible.Shaman.MuradinsLastStand;
+
+	public override string[] RelatedCards => new string[]
+	{
+		HearthDb.CardIds.Collectible.Shaman.MuradinsLastStand,
+	};
+
+	public FriendlyAttacksThisGameCounter(bool controlledByPlayer, GameV2 game) : base(controlledByPlayer, game)
+	{
+	}
+
+	public override bool ShouldShow()
+	{
+		if(!Game.IsTraditionalHearthstoneMatch)
+			return false;
+		return IsPlayerCounter && InPlayerDeckOrKnown(RelatedCards);
+	}
+
+	public override string[] GetCardsToDisplay()
+	{
+		return RelatedCards;
+	}
+
+	public override string ValueToShow() => string.Format(LocUtil.Get("Counter_AnimalCompanionCost"),
+		Math.Max(9 - Counter, 0));
+	public override void HandleTagChange(GameTag tag, IHsGameState gameState, Entity entity, int value, int prevValue)
+	{
+		if(!Game.IsTraditionalHearthstoneMatch)
+			return;
+
+		if(entity.IsControlledBy(Game.Player.Id) != IsPlayerCounter)
+			return;
+
+		if(tag != GameTag.ATTACKING)
+			return;
+
+		if(value == 0)
+			return;
+
+		Counter++;
+	}
+}